client/requirements.txt
kafka-python
textblob
cloudant
